work_captcha_recognition/tfrecord/tfrecord_model_train/requirements.txt
absl-py==0.13.0
astor==0.8.1
cached-property==1.5.2
captcha==0.3
cycler==0.10.0
dataclasses==0.8
gast==0.5.0
google-pasta==0.2.0
grpcio==1.38.1
h5py==3.1.0
importlib-metadata==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
numpy==1.19.5
Pillow==8.3.0
protobuf==3.17.3
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.16.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
